requirements.txt
Django==5.2.5
pillow==11.3.0
werkzeug==3.1.3
whitenoise==6.9.0
django-filter==25.1
python-decouple==3.8
html2text==2025.4.15
drf-extensions==0.8.0
django-extensions==4.1
dj-database-url==3.0.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
drf-standardized-errors[openapi]==0.15.0
djangorestframework_simplejwt==5.5.1
djangorestframework-camel-case==1.4.2
